requirements.txt
# Core ML
scikit-learn>=1.3.0
pycaret[full]>=3.3.0

# Data
pandas>=2.0.0
numpy>=1.24.0

# Model persistence
joblib>=1.3.0

# HTTP client (WeatherAPI)
httpx>=0.27.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Database
supabase>=2.4.0

# Scheduler
APScheduler>=3.10.0

# Environment
python-dotenv>=1.0.0

# Optional: online learning
river>=0.21.0